Move upsert/delete functions to `client-rmw-sql`
